[#217] BusinessRegistry: Make it not possible to revoke an expired or revoked key.
The unit test "Already expired AND revoked pub key" now fails, because it trys to revoke a key that is already expired. It would be possible to fix this test, by setting the expiry time into the future, revoking the key and then waiting for the key to expire, but this is already done in the client test "Test where the key has an expiry time and a revoked time which expired after it was revoked and both revoked and expired time have passed." and so it seems like a waste of effort to fix this test.