[#501] Trails: Remove CORS middleware.
Now that we are going via the EDAPI it adds the CORS headers and it was creating problems that both services were adding CORS headers. We now hand over responsibility for adding the CORS headers to the EDAPI and assume all requests to the Trails service will be proxied by that and have the CORS headers added there.